+++ prequ/scripts/prequ.py
@@ -1,6 +1,7 @@
 import click
 
 from . import build_wheels
+from . import check
 from . import compile_in
 from . import compile
 from . import sync
@@ -16,6 +17,7 @@ def main():
 
 
 main.add_command(build_wheels.main, 'build-wheels')
+main.add_command(check.main, 'check')
 main.add_command(compile_in.cli, 'compile-in')
 main.add_command(compile.main, 'compile')
 main.add_command(sync.cli, 'sync')